updated Exercise 01 as per chapter 11
