+++ Chapter05/Exercise 5.01/src/App.css
@@ -1,0 +1,3 @@
+.App {
+  text-align: left;
+}
